Move tests to OpenRosaServerClientFactory
