+++ collect_app/src/main/java/org/odk/collect/android/http/openrosa/OpenRosaHttpInterface.java
@@ -14,10 +14,15 @@
  * limitations under the License.
  */
 
-package org.odk.collect.android.http;
+package org.odk.collect.android.http.openrosa;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
+import org.odk.collect.android.http.HttpCredentialsInterface;
+import org.odk.collect.android.http.HttpGetResult;
+import org.odk.collect.android.http.HttpHeadResult;
+import org.odk.collect.android.http.HttpPostResult;
 
 import java.io.File;
 import java.io.IOException;
