+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyModule.java
@@ -13,8 +13,9 @@ import org.odk.collect.android.dao.InstancesDao;
 import org.odk.collect.android.events.RxEventBus;
 import org.odk.collect.android.http.CollectServerClient;
 import org.odk.collect.android.http.CollectThenSystemContentTypeMapper;
-import org.odk.collect.android.http.OkHttpConnection;
-import org.odk.collect.android.http.OpenRosaHttpInterface;
+import org.odk.collect.android.http.okhttp.OkHttpConnection;
+import org.odk.collect.android.http.okhttp.OkHttpOpenRosaServerClientFactory;
+import org.odk.collect.android.http.openrosa.OpenRosaHttpInterface;
 import org.odk.collect.android.tasks.sms.SmsSubmissionManager;
 import org.odk.collect.android.tasks.sms.contracts.SmsSubmissionManagerContract;
 import org.odk.collect.android.utilities.DownloadFormListUtils;
@@ -25,6 +26,7 @@ import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
+import okhttp3.OkHttpClient;
 
 /**
  * Add dependency providers here (annotated with @Provides)
@@ -72,11 +74,14 @@ public class AppDependencyModule {
     @Provides
     @Singleton
     OpenRosaHttpInterface provideHttpInterface(MimeTypeMap mimeTypeMap) {
-        return new OkHttpConnection(null, new CollectThenSystemContentTypeMapper(mimeTypeMap));
+        return new OkHttpConnection(
+                new OkHttpOpenRosaServerClientFactory(new OkHttpClient.Builder()),
+                new CollectThenSystemContentTypeMapper(mimeTypeMap)
+        );
     }
 
     @Provides
-    public CollectServerClient provideCollectServerClient(OpenRosaHttpInterface httpInterface, WebCredentialsUtils webCredentialsUtils) {
+    CollectServerClient provideCollectServerClient(OpenRosaHttpInterface httpInterface, WebCredentialsUtils webCredentialsUtils) {
         return new CollectServerClient(httpInterface, webCredentialsUtils);
     }
