+++ collect_app/src/test/java/org/odk/collect/android/http/OkHttpConnectionHeadRequestTest.java
@@ -3,24 +3,21 @@ package org.odk.collect.android.http;
 import android.webkit.MimeTypeMap;
 
 import org.junit.runner.RunWith;
+import org.odk.collect.android.http.okhttp.OkHttpConnection;
+import org.odk.collect.android.http.okhttp.OkHttpOpenRosaServerClientFactory;
+import org.odk.collect.android.http.openrosa.OpenRosaHttpInterface;
 import org.robolectric.RobolectricTestRunner;
 
 import okhttp3.OkHttpClient;
-import okhttp3.tls.internal.TlsUtil;
 
 @RunWith(RobolectricTestRunner.class)
 public class OkHttpConnectionHeadRequestTest extends OpenRosaHeadRequestTest {
 
     @Override
     protected OpenRosaHttpInterface buildSubject() {
-        return new OkHttpConnection(new OkHttpClient.Builder()
-                .sslSocketFactory(TlsUtil.localhost().sslSocketFactory(), TlsUtil.localhost().trustManager()),
+        return new OkHttpConnection(
+                new OkHttpOpenRosaServerClientFactory(new OkHttpClient.Builder()),
                 new CollectThenSystemContentTypeMapper(MimeTypeMap.getSingleton())
         );
-    }
-
-    @Override
-    protected Boolean useRealHttps() {
-        return true;
     }
 }
