+++ collect_app/src/test/java/org/odk/collect/android/http/OkHttpConnectionPostRequest.java
@@ -1,24 +1,21 @@
 package org.odk.collect.android.http;
 
 import org.junit.runner.RunWith;
+import org.odk.collect.android.http.okhttp.OkHttpConnection;
+import org.odk.collect.android.http.okhttp.OkHttpOpenRosaServerClientFactory;
+import org.odk.collect.android.http.openrosa.OpenRosaHttpInterface;
 import org.robolectric.RobolectricTestRunner;
 
 import okhttp3.OkHttpClient;
-import okhttp3.tls.internal.TlsUtil;
 
 @RunWith(RobolectricTestRunner.class)
 public class OkHttpConnectionPostRequest extends OpenRosaPostRequestTest {
 
     @Override
     protected OpenRosaHttpInterface buildSubject(OpenRosaHttpInterface.FileToContentTypeMapper mapper) {
-        return new OkHttpConnection(new OkHttpClient.Builder()
-                .sslSocketFactory(TlsUtil.localhost().sslSocketFactory(), TlsUtil.localhost().trustManager()),
+        return new OkHttpConnection(
+                new OkHttpOpenRosaServerClientFactory(new OkHttpClient.Builder()),
                 mapper
         );
-    }
-
-    @Override
-    protected Boolean useRealHttps() {
-        return true;
     }
 }
