+++ collect_app/src/test/java/org/odk/collect/android/http/OpenRosaGetRequestTest.java
@@ -5,6 +5,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.odk.collect.android.BuildConfig;
+import org.odk.collect.android.http.openrosa.OpenRosaHttpInterface;
 import org.odk.collect.android.utilities.FileUtils;
 
 import java.io.ByteArrayInputStream;
@@ -17,23 +18,18 @@ import java.util.zip.GZIPOutputStream;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
-import okhttp3.tls.internal.TlsUtil;
 import okio.Buffer;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.isEmptyOrNullString;
 import static org.hamcrest.Matchers.nullValue;
-import static org.hamcrest.Matchers.startsWith;
 import static org.junit.Assert.assertThat;
 
 public abstract class OpenRosaGetRequestTest {
 
     protected abstract OpenRosaHttpInterface buildSubject();
 
-    protected abstract Boolean useRealHttps();
-
     private final MockWebServer mockWebServer = new MockWebServer();
-    private MockWebServer httpsMockWebServer;
     private OpenRosaHttpInterface subject;
 
     @Before
@@ -45,11 +41,6 @@ public abstract class OpenRosaGetRequestTest {
     @After
     public void teardown() throws Exception {
         mockWebServer.shutdown();
-
-        if (httpsMockWebServer != null) {
-            httpsMockWebServer.shutdown();
-            httpsMockWebServer = null;
-        }
     }
 
     @Test
@@ -77,152 +68,6 @@ public abstract class OpenRosaGetRequestTest {
                 "null %s/%s",
                 BuildConfig.APPLICATION_ID,
                 BuildConfig.VERSION_NAME)));
-    }
-
-    @Test
-    public void sendsOpenRosaHeaders() throws Exception {
-        mockWebServer.enqueue(new MockResponse());
-
-        subject.executeGetRequest(mockWebServer.url("").uri(), null, null);
-
-        RecordedRequest request = mockWebServer.takeRequest();
-        assertThat(request.getHeader("X-OpenRosa-Version"), equalTo("1.0"));
-    }
-
-    @Test
-    public void sendsAcceptsGzipHeader() throws Exception {
-        mockWebServer.enqueue(new MockResponse());
-
-        subject.executeGetRequest(mockWebServer.url("").uri(), null, null);
-
-        RecordedRequest request = mockWebServer.takeRequest();
-        assertThat(request.getHeader("Accept-Encoding"), equalTo("gzip"));
-    }
-
-    @Test
-    public void withCredentials_whenBasicChallengeReceived_doesNotRetryWithCredentials() throws Exception {
-        mockWebServer.enqueue(new MockResponse()
-                .setResponseCode(401)
-                .addHeader("WWW-Authenticate: Basic realm=\"protected area\"")
-                .setBody("Please authenticate."));
-        mockWebServer.enqueue(new MockResponse());
-
-        subject.executeGetRequest(mockWebServer.url("").uri(), null, new HttpCredentials("user", "pass"));
-
-        assertThat(mockWebServer.getRequestCount(), equalTo(1));
-    }
-
-    @Test
-    public void withCredentials_whenBasicChallengeReceived_whenHttps_retriesWithCredentials() throws Exception {
-        startHttpsMockWebServer();
-
-        httpsMockWebServer.enqueue(new MockResponse()
-                .setResponseCode(401)
-                .addHeader("WWW-Authenticate: Basic realm=\"protected area\"")
-                .setBody("Please authenticate."));
-        httpsMockWebServer.enqueue(new MockResponse());
-
-        subject.executeGetRequest(httpsMockWebServer.url("").uri(), null, new HttpCredentials("user", "pass"));
-
-        assertThat(httpsMockWebServer.getRequestCount(), equalTo(2));
-        httpsMockWebServer.takeRequest();
-        RecordedRequest request = httpsMockWebServer.takeRequest();
-        assertThat(request.getHeader("Authorization"), equalTo("Basic dXNlcjpwYXNz"));
-    }
-
-    @Test
-    public void withCredentials_whenDigestChallengeReceived_retriesWithCredentials() throws Exception {
-        mockWebServer.enqueue(new MockResponse()
-                .setResponseCode(401)
-                .addHeader("WWW-Authenticate: Digest realm=\"ODK Aggregate\", qop=\"auth\", nonce=\"MTU2NTA4MjEzODI4OTpmMjc4MDM5N2YxZTJiNDRiNjNiYTBiMThiOWQ4ZTlkMg==\"")
-                .setBody("Please authenticate."));
-        mockWebServer.enqueue(new MockResponse());
-
-        subject.executeGetRequest(mockWebServer.url("").uri(), null, new HttpCredentials("user", "pass"));
-
-        assertThat(mockWebServer.getRequestCount(), equalTo(2));
-        mockWebServer.takeRequest();
-        RecordedRequest request = mockWebServer.takeRequest();
-        assertThat(request.getHeader("Authorization"), startsWith("Digest"));
-    }
-
-    @Test
-    public void withCredentials_whenDigestChallengeReceived_whenHttps_retriesWithCredentials() throws Exception {
-        startHttpsMockWebServer();
-
-        httpsMockWebServer.enqueue(new MockResponse()
-                .setResponseCode(401)
-                .addHeader("WWW-Authenticate: Digest realm=\"ODK Aggregate\", qop=\"auth\", nonce=\"MTU2NTA4MjEzODI4OTpmMjc4MDM5N2YxZTJiNDRiNjNiYTBiMThiOWQ4ZTlkMg==\"")
-                .setBody("Please authenticate."));
-        httpsMockWebServer.enqueue(new MockResponse());
-
-        subject.executeGetRequest(httpsMockWebServer.url("").uri(), null, new HttpCredentials("user", "pass"));
-
-        assertThat(httpsMockWebServer.getRequestCount(), equalTo(2));
-        httpsMockWebServer.takeRequest();
-        RecordedRequest request = httpsMockWebServer.takeRequest();
-        assertThat(request.getHeader("Authorization"), startsWith("Digest"));
-    }
-
-    @Test
-    public void withCredentials_onceBasicChallenged_whenHttps_proactivelySendsCredentials() throws Exception {
-        startHttpsMockWebServer();
-
-        httpsMockWebServer.enqueue(new MockResponse()
-                .setResponseCode(401)
-                .addHeader("WWW-Authenticate: Basic realm=\"protected area\"")
-                .setBody("Please authenticate."));
-        httpsMockWebServer.enqueue(new MockResponse());
-        httpsMockWebServer.enqueue(new MockResponse());
-
-        subject.executeGetRequest(httpsMockWebServer.url("").uri(), null, new HttpCredentials("user", "pass"));
-        subject.executeGetRequest(httpsMockWebServer.url("/different").uri(), null, new HttpCredentials("user", "pass"));
-
-        assertThat(httpsMockWebServer.getRequestCount(), equalTo(3));
-        httpsMockWebServer.takeRequest();
-        httpsMockWebServer.takeRequest();
-        RecordedRequest request = httpsMockWebServer.takeRequest();
-        assertThat(request.getHeader("Authorization"), equalTo("Basic dXNlcjpwYXNz"));
-    }
-
-    @Test
-    public void withCredentials_onceDigestChallenged_proactivelySendsCredentials() throws Exception {
-        mockWebServer.enqueue(new MockResponse()
-                .setResponseCode(401)
-                .addHeader("WWW-Authenticate: Digest realm=\"ODK Aggregate\", qop=\"auth\", nonce=\"MTU2NTA4MjEzODI4OTpmMjc4MDM5N2YxZTJiNDRiNjNiYTBiMThiOWQ4ZTlkMg==\"")
-                .setBody("Please authenticate."));
-        mockWebServer.enqueue(new MockResponse());
-        mockWebServer.enqueue(new MockResponse());
-
-        subject.executeGetRequest(mockWebServer.url("").uri(), null, new HttpCredentials("user", "pass"));
-        subject.executeGetRequest(mockWebServer.url("/different").uri(), null, new HttpCredentials("user", "pass"));
-
-        assertThat(mockWebServer.getRequestCount(), equalTo(3));
-        mockWebServer.takeRequest();
-        mockWebServer.takeRequest();
-        RecordedRequest request = mockWebServer.takeRequest();
-        assertThat(request.getHeader("Authorization"), startsWith("Digest"));
-    }
-
-    @Test
-    public void withCredentials_onceDigestChallenged_whenHttps_proactivelySendsCredentials() throws Exception {
-        startHttpsMockWebServer();
-
-        httpsMockWebServer.enqueue(new MockResponse()
-                .setResponseCode(401)
-                .addHeader("WWW-Authenticate: Digest realm=\"ODK Aggregate\", qop=\"auth\", nonce=\"MTU2NTA4MjEzODI4OTpmMjc4MDM5N2YxZTJiNDRiNjNiYTBiMThiOWQ4ZTlkMg==\"")
-                .setBody("Please authenticate."));
-        httpsMockWebServer.enqueue(new MockResponse());
-        httpsMockWebServer.enqueue(new MockResponse());
-
-        subject.executeGetRequest(httpsMockWebServer.url("").uri(), null, new HttpCredentials("user", "pass"));
-        subject.executeGetRequest(httpsMockWebServer.url("/different").uri(), null, new HttpCredentials("user", "pass"));
-
-        assertThat(httpsMockWebServer.getRequestCount(), equalTo(3));
-        httpsMockWebServer.takeRequest();
-        httpsMockWebServer.takeRequest();
-        RecordedRequest request = httpsMockWebServer.takeRequest();
-        assertThat(request.getHeader("Authorization"), startsWith("Digest"));
     }
 
     @Test
@@ -324,16 +169,6 @@ public abstract class OpenRosaGetRequestTest {
         HttpGetResult result2 = subject.executeGetRequest(mockWebServer.url("").uri(), null, null);
         assertThat(result2.getInputStream(), nullValue());
         assertThat(result2.getStatusCode(), equalTo(304));
-    }
-
-    private void startHttpsMockWebServer() throws IOException {
-        httpsMockWebServer = new MockWebServer();
-
-        if (useRealHttps()) {
-            httpsMockWebServer.useHttps(TlsUtil.localhost().sslSocketFactory(), false);
-        }
-
-        httpsMockWebServer.start(8443);
     }
 
     private static byte[] gzip(String data) throws IOException {
