+++ collect_app/src/test/java/org/odk/collect/android/http/OpenRosaHeadRequestTest.java
@@ -4,14 +4,13 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.odk.collect.android.BuildConfig;
+import org.odk.collect.android.http.openrosa.OpenRosaHttpInterface;
 
-import java.io.IOException;
 import java.net.URI;
 
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
-import okhttp3.tls.internal.TlsUtil;
 
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.Matchers.equalTo;
@@ -24,10 +23,7 @@ public abstract class OpenRosaHeadRequestTest {
 
     protected abstract OpenRosaHttpInterface buildSubject();
 
-    protected abstract Boolean useRealHttps();
-
     private final MockWebServer mockWebServer = new MockWebServer();
-    private MockWebServer httpsMockWebServer;
     private OpenRosaHttpInterface subject;
 
     @Before
@@ -39,11 +35,6 @@ public abstract class OpenRosaHeadRequestTest {
     @After
     public void teardown() throws Exception {
         mockWebServer.shutdown();
-
-        if (httpsMockWebServer != null) {
-            httpsMockWebServer.shutdown();
-            httpsMockWebServer = null;
-        }
     }
 
     @Test
@@ -74,16 +65,6 @@ public abstract class OpenRosaHeadRequestTest {
     }
 
     @Test
-    public void sendsOpenRosaHeaders() throws Exception {
-        mockWebServer.enqueue(new MockResponse());
-
-        subject.executeHeadRequest(mockWebServer.url("").uri(), null);
-
-        RecordedRequest request = mockWebServer.takeRequest();
-        assertThat(request.getHeader("X-OpenRosa-Version"), equalTo("1.0"));
-    }
-
-    @Test
     public void whenLastRequestSetCookies_nextRequestDoesNotSendThem() throws Exception {
         mockWebServer.enqueue(new MockResponse()
                 .addHeader("Set-Cookie", "blah=blah"));
@@ -110,35 +91,6 @@ public abstract class OpenRosaHeadRequestTest {
     }
 
     @Test
-    public void withCredentials_whenHttp_doesNotRetryWithCredentials() throws Exception {
-        mockWebServer.enqueue(new MockResponse()
-                .setResponseCode(401)
-                .addHeader("WWW-Authenticate: Basic realm=\"protected area\""));
-        mockWebServer.enqueue(new MockResponse());
-
-        subject.executeHeadRequest(mockWebServer.url("").uri(), new HttpCredentials("user", "pass"));
-
-        assertThat(mockWebServer.getRequestCount(), equalTo(1));
-    }
-
-    @Test
-    public void withCredentials_whenHttps_retriesWithCredentials() throws Exception  {
-        startHttpsMockWebServer();
-
-        httpsMockWebServer.enqueue(new MockResponse()
-                .setResponseCode(401)
-                .addHeader("WWW-Authenticate: Basic realm=\"protected area\""));
-        httpsMockWebServer.enqueue(new MockResponse());
-
-        buildSubject().executeHeadRequest(httpsMockWebServer.url("").uri(), new HttpCredentials("user", "pass"));
-
-        assertThat(httpsMockWebServer.getRequestCount(), equalTo(2));
-        httpsMockWebServer.takeRequest();
-        RecordedRequest request = httpsMockWebServer.takeRequest();
-        assertThat(request.getHeader("Authorization"), equalTo("Basic dXNlcjpwYXNz"));
-    }
-
-    @Test
     public void whenRequestFails_throwsExceptionWithMessage() {
         try {
             subject.executeHeadRequest(new URI("http://localhost:8443"), null);
@@ -146,15 +98,5 @@ public abstract class OpenRosaHeadRequestTest {
         } catch (Exception e) {
             assertThat(e.getMessage(), not(isEmptyString()));
         }
-    }
-
-    private void startHttpsMockWebServer() throws IOException {
-        httpsMockWebServer = new MockWebServer();
-
-        if (useRealHttps()) {
-            httpsMockWebServer.useHttps(TlsUtil.localhost().sslSocketFactory(), false);
-        }
-
-        httpsMockWebServer.start(8443);
     }
 }
