+++ collect_app/src/test/java/org/odk/collect/android/http/OpenRosaPostRequestTest.java
@@ -5,6 +5,7 @@ import androidx.annotation.NonNull;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.odk.collect.android.http.openrosa.OpenRosaHttpInterface;
 
 import java.io.BufferedWriter;
 import java.io.ByteArrayOutputStream;
@@ -21,7 +22,6 @@ import java.util.zip.GZIPOutputStream;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
-import okhttp3.tls.internal.TlsUtil;
 import okio.Buffer;
 
 import static java.util.Arrays.asList;
@@ -39,10 +39,7 @@ public abstract class OpenRosaPostRequestTest {
 
     protected abstract OpenRosaHttpInterface buildSubject(OpenRosaHttpInterface.FileToContentTypeMapper mapper);
 
-    protected abstract Boolean useRealHttps();
-
     private final MockWebServer mockWebServer = new MockWebServer();
-    private MockWebServer httpsMockWebServer;
     private OpenRosaHttpInterface subject;
 
     @Before
@@ -54,11 +51,6 @@ public abstract class OpenRosaPostRequestTest {
     @After
     public void teardown() throws Exception {
         mockWebServer.shutdown();
-
-        if (httpsMockWebServer != null) {
-            httpsMockWebServer.shutdown();
-            httpsMockWebServer = null;
-        }
     }
 
     @Test
@@ -73,39 +65,6 @@ public abstract class OpenRosaPostRequestTest {
         RecordedRequest request = mockWebServer.takeRequest();
         assertThat(request.getMethod(), equalTo("POST"));
         assertThat(request.getRequestUrl().uri(), equalTo(uri));
-    }
-
-    @Test
-    public void withCredentials_whenHttp_doesNotRetryWithCredentials() throws Exception {
-        mockWebServer.enqueue(new MockResponse()
-                .setResponseCode(401)
-                .addHeader("WWW-Authenticate: Basic realm=\"protected area\"")
-                .setBody("Please authenticate."));
-        mockWebServer.enqueue(new MockResponse());
-
-        URI uri = mockWebServer.url("").uri();
-        subject.uploadSubmissionFile(new ArrayList<>(), File.createTempFile("blah", "blah"), uri, new HttpCredentials("user", "pass"), 0);
-
-        assertThat(mockWebServer.getRequestCount(), equalTo(1));
-    }
-
-    @Test
-    public void withCredentials_whenHttps_retriesWithCredentials() throws Exception {
-        startHttpsMockWebServer();
-
-        httpsMockWebServer.enqueue(new MockResponse()
-                .setResponseCode(401)
-                .addHeader("WWW-Authenticate: Basic realm=\"protected area\"")
-                .setBody("Please authenticate."));
-        httpsMockWebServer.enqueue(new MockResponse());
-
-        URI uri = httpsMockWebServer.url("").uri();
-        subject.uploadSubmissionFile(new ArrayList<>(), File.createTempFile("blah", "blah"), uri, new HttpCredentials("user", "pass"), 0);
-
-        assertThat(httpsMockWebServer.getRequestCount(), equalTo(2));
-        httpsMockWebServer.takeRequest();
-        RecordedRequest request = httpsMockWebServer.takeRequest();
-        assertThat(request.getHeader("Authorization"), equalTo("Basic dXNlcjpwYXNz"));
     }
 
     @Test
@@ -300,16 +259,6 @@ public abstract class OpenRosaPostRequestTest {
         assertThat(mockWebServer.getRequestCount(), equalTo(2));
         assertThat(response, notNullValue());
         assertThat(response.getResponseCode(), equalTo(500));
-    }
-
-    private void startHttpsMockWebServer() throws IOException {
-        httpsMockWebServer = new MockWebServer();
-
-        if (useRealHttps()) {
-            httpsMockWebServer.useHttps(TlsUtil.localhost().sslSocketFactory(), false);
-        }
-
-        httpsMockWebServer.start(8443);
     }
 
     private File createTempFile(String content) throws Exception {
